parse version check and updated file
